Add files via upload
A semi-abstract piece using coding principles from the book and from Python.org. code took me about 17 hours to do as I kept messing up and not getting the desired result. I had to look up on Python.org how to fix my many mistakes. This included some forgetting of () which python wasn't giving me error for but ran without. also looping numbers N=100 had to go outside of #for i in range(): and the brackets for lists of data.