Mariana.rings.Tokyo
An explanation of my Turtle Art lesson plan and the Turtle art project I made. the art is a semi-abstract piece using coding principles from the book and from Python.org. The code took me about 17 hours to do as I kept messing up and not getting the desired result. I had to look up on Python.org how to fix my many mistakes. This included some forgetting of () which python wasn't giving me error for but ran without the code. also looping numbers N=10 had to go outside of #for i in range(): and the brackets for lists of data.